Fix JSON output with missing package license text.
The bucket objects contain the license text for the packages in the bucket. The
other generators use the bucket text. Due to the hashing in the id of the
bucket map, all packages in the bucket have the same text. Some packages don't
have text and the default text is looked up in and added to the bucket, but not
the packages. The JSON generator was not using the bucket text. Also ran
lint:fix.